requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
click==8.3.0
fastapi==0.116.2
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.4
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0
